break out change - do some cleanup
