+++ automerge-backend/tests/apply_local_change.rs
@@ -50,7 +50,7 @@ fn test_apply_local_change() {
             insert: false,
         }],
     }
-    .into_bin();
+    .encode();
     assert_eq!(changes[0], &expected_change);
 
     let expected_patch = Patch {
@@ -183,7 +183,7 @@ fn test_handle_concurrent_frontend_and_backend_changes() {
             insert: false,
         }],
     }
-    .into_bin();
+    .encode();
 
     let mut expected_change1 = Change {
         actor_id: actor.clone(),
@@ -252,9 +252,9 @@ fn test_handle_concurrent_frontend_and_backend_changes() {
     expected_change3.time = change23.time;
     expected_change3.deps = vec![change01.hash];
 
-    assert_eq!(change01, &&expected_change1.into_bin());
-    assert_eq!(change12, &expected_change2.into_bin());
-    assert_eq!(change23, &&expected_change3.into_bin());
+    assert_eq!(change01, &&expected_change1.encode());
+    assert_eq!(change12, &expected_change2.encode());
+    assert_eq!(change23, &&expected_change3.encode());
 }
 
 #[test]
@@ -276,7 +276,7 @@ fn test_transform_list_indexes_into_element_ids() {
             insert: false,
         }],
     }
-    .to_bin();
+    .encode();
 
     let remote2 = Change {
         actor_id: remote_actor,
@@ -293,7 +293,7 @@ fn test_transform_list_indexes_into_element_ids() {
             pred: Vec::new(),
         }],
     }
-    .into_bin();
+    .encode();
 
     let local1 = ChangeRequest {
         actor: actor.clone(),
@@ -441,7 +441,7 @@ fn test_transform_list_indexes_into_element_ids() {
 
     backend.apply_local_change(local3).unwrap();
     let changes3 = backend.get_changes(&[remote2.hash, change23.hash]);
-    let change34 = changes3.get(0).unwrap().to_change();
+    let change34 = changes3.get(0).unwrap().decode();
 
     expected_change1.time = change12.time;
     expected_change2.time = change23.time;
@@ -449,8 +449,8 @@ fn test_transform_list_indexes_into_element_ids() {
     expected_change3.time = change34.time;
     expected_change3.deps = vec![remote2.hash, change23.hash];
 
-    assert_eq!(change12, &expected_change1.into_bin());
-    assert_eq!(change23, &expected_change2.into_bin());
+    assert_eq!(change12, &expected_change1.encode());
+    assert_eq!(change23, &expected_change2.encode());
     assert_changes_equal(change34, expected_change3);
 }
 
@@ -563,7 +563,7 @@ fn test_handle_list_insertion_and_deletion_in_same_change() {
             pred: Vec::new(),
         }],
     }
-    .into_bin();
+    .encode();
 
     let expected_change2 = Change {
         actor_id: actor,
@@ -591,7 +591,7 @@ fn test_handle_list_insertion_and_deletion_in_same_change() {
             },
         ],
     }
-    .into_bin();
+    .encode();
 
     assert_eq!(change1, expected_change1);
     assert_eq!(change2, expected_change2);
