+++ automerge-c/src/lib.rs
@@ -3,7 +3,7 @@ extern crate errno;
 extern crate libc;
 extern crate serde;
 
-use automerge_backend::{ BinChange, AutomergeError };
+use automerge_backend::{AutomergeError, BinChange};
 use automerge_protocol::{ActorID, ChangeRequest};
 use errno::{set_errno, Errno};
 use serde::ser::Serialize;
